Refactor code and add retryable error
