+++ api/queue/worker.go
@@ -1,6 +1,7 @@
 package queue
 
 import (
+	"sync"
 	"time"
 
 	"github.com/gojek/merlin/log"
@@ -10,11 +11,12 @@ import (
 var (
 	findJobQuery           = "SELECT * FROM jobs where id = ? AND completed = ? FOR UPDATE SKIP LOCKED"
 	allIncompleteJobsQuery = "SELECT * FROM jobs where completed = ? FOR UPDATE SKIP LOCKED"
+	delayOfRetry           = 5 * time.Second
 )
 
 type worker struct {
 	quitChan   chan bool
-	jobFuncMap map[string]JobFn
+	jobFuncMap *sync.Map
 	jobChan    chan *Job
 	db         *gorm.DB
 }
@@ -24,7 +26,7 @@ func newWorker(db *gorm.DB, jobChan chan *Job) *worker {
 	return &worker{db: db, jobChan: jobChan, quitChan: quitChan}
 }
 
-func (w *worker) updateWorkerJobFunction(jobFn map[string]JobFn) {
+func (w *worker) updateWorkerJobFunction(jobFn *sync.Map) {
 	w.jobFuncMap = jobFn
 }
 
@@ -47,23 +49,36 @@ func (w *worker) processJob(job *Job) {
 	err := tx.Raw(findJobQuery, job.ID, false).Scan(&refreshedJob).Error
 	if err == gorm.ErrRecordNotFound {
 		tx.Rollback()
-		log.Warnf("Job records not found with id: %d", job.ID)
+		log.Warnf("Job with ID:%d is still running in other process or already ran successfully", job.ID)
 		return
 	}
 
 	if err != nil {
 		tx.Rollback()
+		w.requeueJob(job)
 		return
 	}
 
-	jobFn, ok := w.jobFuncMap[job.Name]
+	fn, ok := w.jobFuncMap.Load(job.Name)
 	if !ok {
 		log.Warnf("There is no function be run for job %s", job.Name)
 		tx.Rollback()
 		return
 	}
+
+	jobFn, ok := fn.(JobFn)
+	if !ok {
+		log.Warnf("Registered function is not correct")
+		tx.Rollback()
+		return
+	}
 	if err := jobFn(job); err != nil {
 		log.Errorf("Job execution is failed, with id:%d and error: %v", job.ID, err)
+		switch err.(type) {
+		case RetryableError:
+			w.requeueJob(job)
+			return
+		}
 	}
 
 	refreshedJob.UpdatedAt = time.Now()
@@ -71,10 +86,19 @@ func (w *worker) processJob(job *Job) {
 	if err := tx.Save(refreshedJob).Error; err != nil {
 		log.Errorf("Failed to save job %d with error: %v", refreshedJob.ID, err)
 		tx.Rollback()
+
+		w.requeueJob(job)
 		return
 	}
 
 	tx.Commit()
+}
+
+func (w *worker) requeueJob(job *Job) {
+	time.Sleep(delayOfRetry) // sleep for specific amount of time before retry
+	go func() {
+		w.jobChan <- job
+	}()
 }
 
 func (w *worker) stop() {
