+++ api/service/prediction_job_service_test.go
@@ -23,7 +23,6 @@ import (
 	"github.com/jinzhu/copier"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/mock"
-	"github.com/stretchr/testify/require"
 	"k8s.io/apimachinery/pkg/util/clock"
 
 	"github.com/gojek/merlin/batch"
@@ -31,7 +30,6 @@ import (
 	imageBuilderMock "github.com/gojek/merlin/imagebuilder/mocks"
 	"github.com/gojek/merlin/mlp"
 	"github.com/gojek/merlin/models"
-	"github.com/gojek/merlin/queue"
 	queueMock "github.com/gojek/merlin/queue/mocks"
 	storageMock "github.com/gojek/merlin/storage/mocks"
 )
@@ -123,90 +121,6 @@ var (
 	}
 )
 
-func TestPrediction_ExecuteDeployment(t *testing.T) {
-	savedJob := new(models.PredictionJob)
-	err := copier.Copy(savedJob, job)
-	require.NoError(t, err)
-
-	savedJob.Config.ImageRef = imageRef
-	testCases := []struct {
-		desc             string
-		deployErr        error
-		controllerMock   func(*mocks.Controller)
-		imageBuilderMock func(*imageBuilderMock.ImageBuilder)
-		mockStorage      func(*storageMock.PredictionJobStorage)
-	}{
-		{
-			desc:      "Success",
-			deployErr: nil,
-			controllerMock: func(ctrl *mocks.Controller) {
-				ctrl.On("Submit", savedJob, project.Name).Return(nil)
-			},
-			imageBuilderMock: func(imgBuilder *imageBuilderMock.ImageBuilder) {
-				imgBuilder.On("BuildImage", project, model, version).Return(imageRef, nil)
-			},
-			mockStorage: func(st *storageMock.PredictionJobStorage) {
-				st.On("Get", savedJob.ID).Return(savedJob, nil)
-			},
-		},
-		{
-			desc:           "Failed: building image fail",
-			deployErr:      fmt.Errorf("failed building image"),
-			controllerMock: func(ctrl *mocks.Controller) {},
-			imageBuilderMock: func(imgBuilder *imageBuilderMock.ImageBuilder) {
-				imgBuilder.On("BuildImage", project, model, version).Return("", fmt.Errorf("failed building image"))
-			},
-			mockStorage: func(st *storageMock.PredictionJobStorage) {
-				st.On("Get", savedJob.ID).Return(savedJob, nil)
-				st.On("Save", savedJob).Return(nil)
-			},
-		},
-		{
-			desc:      "Failed: submit job failed",
-			deployErr: fmt.Errorf("failed submit job"),
-			controllerMock: func(ctrl *mocks.Controller) {
-				ctrl.On("Submit", savedJob, project.Name).Return(fmt.Errorf("failed submit job"))
-			},
-			imageBuilderMock: func(imgBuilder *imageBuilderMock.ImageBuilder) {
-				imgBuilder.On("BuildImage", project, model, version).Return(imageRef, nil)
-			},
-			mockStorage: func(st *storageMock.PredictionJobStorage) {
-				st.On("Get", savedJob.ID).Return(savedJob, nil)
-				st.On("Save", savedJob).Return(nil)
-			},
-		},
-	}
-	for _, tC := range testCases {
-		t.Run(tC.desc, func(t *testing.T) {
-			svc, mockControllers, mockImageBuilder, mockStorage, _ := newMockPredictionJobService()
-			mockController := mockControllers[envName]
-			ctrl := mockController.(*mocks.Controller)
-			tC.controllerMock(ctrl)
-			tC.imageBuilderMock(mockImageBuilder)
-			tC.mockStorage(mockStorage)
-			job := &queue.Job{
-				Name: "job",
-				Arguments: queue.Argument{
-					dataArgKey: &BatchJobArgs{
-						Job:         savedJob,
-						Model:       model,
-						Version:     version,
-						Project:     project,
-						Environment: predJobEnv,
-					},
-				},
-			}
-
-			err := svc.ExecuteDeployment(job)
-			if tC.deployErr != nil {
-				assert.Equal(t, tC.deployErr, err)
-			}
-			mockStorage.AssertExpectations(t)
-			mockImageBuilder.AssertExpectations(t)
-			mockController.(*mocks.Controller).AssertExpectations(t)
-		})
-	}
-}
 func TestGetPredictionJob(t *testing.T) {
 	svc, _, _, mockStorage, _ := newMockPredictionJobService()
 	mockStorage.On("Get", job.ID).Return(job, nil)
